+++ agent/consul/server.go
@@ -1455,74 +1455,6 @@ func (s *Server) isReadyForConsistentReads() bool {
 	return atomic.LoadInt32(&s.readyForConsistentReads) == 1
 }
 
-// CreateACLToken will create an ACL token from the given template
-// TODO: move to autoConfigBackend
-func (s *Server) CreateACLToken(template *structs.ACLToken) (*structs.ACLToken, error) {
-	// we have to require local tokens or else it would require having these servers use a token with acl:write to make a
-	// token create RPC to the servers in the primary DC.
-	if !s.LocalTokensEnabled() {
-		return nil, fmt.Errorf("Agent Auto Configuration requires local token usage to be enabled in this datacenter: %s", s.config.Datacenter)
-	}
-
-	newToken := *template
-
-	// generate the accessor id
-	if newToken.AccessorID == "" {
-		accessor, err := lib.GenerateUUID(s.checkTokenUUID)
-		if err != nil {
-			return nil, err
-		}
-
-		newToken.AccessorID = accessor
-	}
-
-	// generate the secret id
-	if newToken.SecretID == "" {
-		secret, err := lib.GenerateUUID(s.checkTokenUUID)
-		if err != nil {
-			return nil, err
-		}
-
-		newToken.SecretID = secret
-	}
-
-	newToken.CreateTime = time.Now()
-
-	req := structs.ACLTokenBatchSetRequest{
-		Tokens: structs.ACLTokens{&newToken},
-		CAS:    false,
-	}
-
-	// perform the request to mint the new token
-	if _, err := s.raftApplyMsgpack(structs.ACLTokenSetRequestType, &req); err != nil {
-		return nil, err
-	}
-
-	// return the full token definition from the FSM
-	_, token, err := s.fsm.State().ACLTokenGetByAccessor(nil, newToken.AccessorID, &newToken.EnterpriseMeta)
-	return token, err
-}
-
-// DatacenterJoinAddresses will return all the strings suitable for usage in
-// retry join operations to connect to the the LAN or LAN segment gossip pool.
-// TODO: move to autoConfigBackend
-func (s *Server) DatacenterJoinAddresses(segment string) ([]string, error) {
-	members, err := s.LANSegmentMembers(segment)
-	if err != nil {
-		return nil, fmt.Errorf("Failed to retrieve members for segment %s - %w", segment, err)
-	}
-
-	var joinAddrs []string
-	for _, m := range members {
-		if ok, _ := metadata.IsConsulServer(m); ok {
-			serfAddr := net.TCPAddr{IP: m.Addr, Port: int(m.Port)}
-			joinAddrs = append(joinAddrs, serfAddr.String())
-		}
-	}
-
-	return joinAddrs, nil
-}
-
 // peersInfoContent is used to help operators understand what happened to the
 // peers.json file. This is written to a file called peers.info in the same
 // location.
